3.27 add obstacle and collision check

+++ src/generate_dynamic_scene/obstacle.cpp
@@ -2,7 +2,7 @@
 
 namespace hdi_plan {
 
-Obstacle::Obstacle(std::string name, Obstacle_type type, bool operation, float size, Eigen::Vector3d position)
+Obstacle::Obstacle(std::string name, Obstacle_type type, bool operation, double size, Eigen::Vector3d position)
 	: name_(name), type_(type), operation_(operation), size_(size), position_(position) {
 
 }
